Move options into its own file
